Add support for exporting to duckdb (via parquet)
This patch adds functionality to export directly to a DuckDB database
via the --duckdb flag or using a ".duckdb" or ".db" file. Optionally one
can change the table name in which data will be imported.

Documentation was mostly copied from existing functions but doctests
were updated and checked for consistency with the results.

Closes kraina-ai/quackosm#94