Add support for exporting to duckdb (via parquet) (#157)
* Add support for exporting to duckdb (via parquet)

This patch adds functionality to export directly to a DuckDB database
via the --duckdb flag or using a ".duckdb" or ".db" file. Optionally one
can change the table name in which data will be imported.

Documentation was mostly copied from existing functions but doctests
were updated and checked for consistency with the results.

Closes kraina-ai/quackosm#94

* Fix open mypy issues

* Fix doctest errors; fix filter_osm_ids typing properly

This patch fixes remaining doctest errors that occured during
kraina-ai/quackosm#157. Meanwhile, a remaining bug was discovered around
the typing of filter_osm_ids. It was solved, too.

* Resolve refurb: immut. tuple over list for suffix

* Fix typo in test_cli.py

* Fix test_cli.py for duckdb: correct file, split args

* chore: add new test case for increased coverage

* chore: add author to the changelog

* feat: add automatic directory generation for duckdb export

* chore: refactor cli codebase

* fix: remove commas

* chore: change working directory test

---------

Co-authored-by: Kamil Raczycki <raczyckikamil@gmail.com>